requirements.txt
anyio==4.3.0
asgiref==3.7.2
astroid==2.11.3
certifi==2024.2.2
colorama==0.4.4
dill==0.3.4
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
Faker==24.1.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pygame==2.1.2
PyJWT==2.8.0
pylint==2.13.7
python-dateutil==2.9.0.post0
pytz==2024.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tomli==2.0.1
toposort==1.10
typing_extensions==4.10.0
tzdata==2024.1
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
wrapt==1.14.0
